Add public profile page (#105)
* add public profile page

* remove unsused styling

* cleaned up styling

* fix public profile page styling

* cleaned up styling

* added label to loading page in public profile page

* fix public api stuffs
