+++ src/models/card.ts
@@ -1,6 +1,12 @@
 import { v4 as uuidv4 } from 'uuid';
+import { ensureHttps } from '@/helpers/api';
+import { asUtcDate } from '@/helpers/time';
 import { isBlankString } from '@/helpers/validator';
 import { CardContent, createNewCardContent } from './card-content';
+import { LazyAudio } from './lazy-audio';
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
 export interface Card {
   id?: number;
   position?: number;
@@ -27,4 +33,35 @@ export function createNewCard(): Card {
 
 export function filterBlankCards(card: Card) {
   return !isBlankString(card.front.text) && !isBlankString(card.back.text);
+}
+
+export function cardToJson(card: Card, deckId?: number) {
+  return {
+    frontText: card.front.text,
+    backText: card.back.text,
+    frontLang: card.front.language,
+    backLang: card.back.language,
+    cardId: card.id, // required when updating cards
+    deckId,
+  };
+}
+
+export function JsonToCard(obj: any): Card {
+  const card = createNewCard();
+  card.id = obj['id'];
+  card.score = obj['score'];
+  card.dateCreated = asUtcDate(obj['date_created']);
+  card.dateUpdated = asUtcDate(obj['date_updated']);
+  card.dateTouched = asUtcDate(obj['date_touched']);
+  card.front = {
+    language: obj['flang'],
+    text: obj['ftext'],
+    audio: new LazyAudio(ensureHttps(obj['faud'])),
+  };
+  card.back = {
+    language: obj['blang'],
+    text: obj['btext'],
+    audio: new LazyAudio(ensureHttps(obj['baud'])),
+  };
+  return card;
 }
